Added search bar to each place and fixed article retrieval
Added the search bar component to each component that needs it.  Back button does not work as expected due to data being changed.  Also fixed a bug so that each time you search for a different company stock, the articles for that company are retreived.  Minor issues when going between components with the Info component, but I think that is mostly due to small backend errors or not properly updating the store on a search.